Clicking on hours in the TimePIckerBaloon sets the text of today's cell for the project with the most hours TimePickerBaloon now only shows up when time is missing for today

+++ DeltekReminder.Lib/TimesheetEntry.cs
@@ -8,7 +8,7 @@ namespace DeltekReminder.Lib
         
         public DateTime Date { get; set; }
         
-        public double Hours { get; set; }
+        public decimal Hours { get; set; }
 
         public int Column { get; set; }
     }
